Write some basic documentation files.
Add content to the README as well as some more detailed documentation
about the game state and RPC interface.

For now, all game state and RPC methods that are implemented are also
documented.  Not-yet-implemented things like specific RPC methods or
the move format are not yet in the docs either.