Define protocol buffer for auth data.
This adds a new top-level directory auth, which contains the xidauth
library.  This will hold all stuff related directly to the authentication
protocol, in particular parsing of passwords and the construction of
the authentication message.

For now, it defines just the protocol buffer that holds the main data
(as encoded by the password).