Add authwithwallet RPC method.
The new method constructs authentication credentials and signs them
directly with the wallet of the attached Xaya Core daemon.

This includes code to actually allow enabling the Xaya Core wallet for
the xid RPC server, so that we have a uniform framework for exposing
the wallet to RPC methods and also for ensuring this is only done when
explicitly requested.