Add crypto addresses to game state.
This adds a new database table to the game state, which can hold
associations of names and crypto identifiers (e.g. "btc" or "eth") to
addresses for that coin/token.

The data is included already in the game state JSON as well, but can not
yet be set through moves.