Implement xid-light binary.
The new "xid-light" binary provides a JSON-RPC server with similar
interface to xid, but runs without needing a Xaya Core connection or
keeping track of the game state.

For now, it just exposes getauthmessage and setauthsignature (which are
both "pure" functions).  In the future, it will also support "getnamestate"
via querying a provided REST endpoint.