Integration test the light mode.
Add an integration test for xid-light.  It mainly checks that all
(requesting null state, name state and constructing and auth message)
works when connected to a local GSP, but it also checks how invalid
server responses are handled and how the connection to the seeder.xaya.io
official REST endpoint (with TLS) works.