Add Docker Compose configuration example.
This adds a new configuration file for Docker Compose, which runs both
Xaya Core and a connected XID instance (exposing its JSON-RPC and REST
interface publicly).

This is a test that this can be done, as well as an example and template
for how production environments could be set up with Docker Compose and
XID (and perhaps additional other GSPs).