Use alpine as base for Docker image.
The xaya/libxayagame Docker image is now built based on alpine rather
than Debian, so make sure xid builds on alpine as well with the
Dockerfile.

Also use alpine as base for xaya/xid, rather than scratch.  This does not
add much size (in fact with the build on alpine, it is smaller now than
the scratch image built on Debian).  With it, basic utilities like
a shell are available on the image as well.