Handle odd-length hex encoded names properly.
When a name is hex-encoded for XMPP, the hex part should have an
even length.  Enforce this explicitly, and add a unit test for it.

Previously such a name would also fail to decode, but it would throw
an exception through codecs.decode.  It is nicer to handle
this gracefully instead.