Move xid-light logic to library.
Factor out most of the xid-light logic from the main source file
to a new "light" file that is part of the xid library, so that it
can be used to start xid-light instances as part of other processes
as well.

The main file now is really just about parsing and verifying flags,
and then passing them on to the new LightInstance class.