Introduce --(no)unsafe_rpc argument.
The new --(no)unsafe_rpc argument (on by default) allows to turn off
'unsafe' RPC methods (which should not be exposed publicly) on the
RPC server.  For now, only 'stop' and 'getcurrentstate' are considered such
an unsafe method (since the wallet-enabled methods are guarded by
--allow_wallet already anyway).

With this, we will be able to open up a public JSON-RPC server for xid
in addition to just the limited REST interface available already for
light-mode clients.