Introduce protocol field for auth credentials.
Add a new protocol field with enum value to the credentials proto that
gets encoded into a password.  We will support different protocols in
the future (in particular, EIP-712 signatures based on move delegation
permissions instead of the Xid GSP gamestate).

The xid binary with verifyauth only supports the old (and default)
protocol.