Add Authenticator's to ejabberd_xidauth.
This is a refactor of the ejabberd_xidauth script to decouple the
individual server definitions from the XID GSP authentication.  Instead
we use an Authenticator abstract class, and each server can have more
than one authenticator defined.

For now, we define exactly the XID GSP one as previously.  In the future
we can add others as well if configured, such as delegation contract.